Non-system tests
The current integration test is more of a system test, and while unit
tests aren’t really possible for the hadoop proxies, there needs to be
some cheaper integration tests. As is, the integration tests are
difficult to extend and since they take ages to run, it is not at all
rewarding to do so.